requirements.txt
streamlit>=1.26.0
numpy>=1.24.0
matplotlib>=3.7.0
google-genai
sounddevice
scipy
pydub